backend/requirements.txt
aiofiles==23.1.0
aiogram==3.0.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.1.0
contourpy==1.1.0
cryptography==41.0.5
cycler==0.11.0
defusedxml==0.8.0rc2
distlib==0.3.6
Django==5.0.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
filelock==3.12.1
fonttools==4.41.1
frozenlist==1.3.3
idna==3.4
iniconfig==2.0.0
kiwisolver==1.4.4
magic-filter==1.0.11
matplotlib==3.7.2
mpmath==1.3.0
multidict==6.0.4
numpy==1.25.1
oauthlib==3.2.2
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
platformdirs==3.5.3
pluggy==1.2.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
PyMySQL==1.1.0
pyparsing==3.0.9
pyTelegramBotAPI==4.12.0
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-vlc==3.0.20123
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.4.2
soupsieve==2.4.1
sqlparse==0.4.4
sympy==1.12
telegram==0.0.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
virtualenv==20.23.0
yarl==1.9.2
